+++ lib/deno-util.js
@@ -1,5 +1,6 @@
 import { readLines } from "https://deno.land/std@0.129.0/io/bufio.ts";
 import jsdom from "https://jspm.dev/npm:jsdom-deno@19.0.1";
+import { patchHTML } from "./page-util.js";
 
 const { JSDOM } = jsdom;
 
@@ -22,19 +23,4 @@ async function parseDocument(html) {
         virtualConsole
     });
     return window.document;
-}
-
-function patchHTML(html) {
-    const REGEXP_HEAD_TAG = /<head>/gi;
-    const REGEXP_SET_IMG_SRC = /function _setImagesSrc/gi;
-    const REPLACEMENT_SET_IMG_SRC = "window._setImagesSrc = _setImagesSrc; function _setImagesSrc";
-    const INJECTED_SCRIPT = `<script>(${emulateLargeScreen.toString()})()</script>`;
-    const REPLACEMENT_HEAD_TAG = `<head>${INJECTED_SCRIPT}`;
-    return html
-        .replace(REGEXP_HEAD_TAG, REPLACEMENT_HEAD_TAG)
-        .replaceAll(REGEXP_SET_IMG_SRC, REPLACEMENT_SET_IMG_SRC); // make sure _setImagesSrc is global (jsdom bug)
-}
-
-function emulateLargeScreen() {
-    window.document.documentElement.__defineGetter__("clientWidth", () => Infinity)
 }
